feat: add currency formatting and update paycheck amounts to use integers

+++ surf-playtime-core/src/main/kotlin/dev/slne/surf/playtime/core/config/PlaytimeConfig.kt
@@ -11,6 +11,6 @@ data class PlaytimeConfig(
 data class PaycheckConfig(
     val enabled: Boolean = false,
     val intervalMinutes: Long = 60,
-    val amount: Double = 250.0,
-    val maxBalance: Double = 100000.0
+    val amount: Int = 250,
+    val maxBalance: Int = 100000
 )
